On branch main Your branch is up to date with 'origin/main'.
Added cpu_serial
This retrieves the serial number of the cpu (currently, this is only
tested to work in an rpi400, this MIGHT be compatible with other rpi
models)

this fetches the CPU ID from:
/sys/firmware/devicetree/base/serial-number

Some basic file validation exists to check if this file exists.

The purpose of this commit is to provide functionality that a
temperature measurement can have a UUID attached to it. Prefferentiably
from some variable or combination of variables that is garunteed
"""Unique"""

Additionally a .gitignore has been created as a QOL for the repo